Refactor the snake_casing map into a common utility
